requirements.txt
Click==7.0
Flask==1.1.1
Jinja2==2.10.1
MarkupSafe==1.1.1
PyJWT==1.7.1
pymongo==3.8.0
Werkzeug==0.15.5
Flask-Cors==3.0.8
bcrypt==3.1.7
uWSGI==2.0.18
